requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
lxml==4.9.3

# Data Processing
pandas==2.1.3
numpy==1.26.2

# NLP
spacy==3.7.2
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.1

# Database
pymongo==4.6.0
motor==3.3.2

# ETL & Orchestration
apache-airflow==2.7.3

# ML Tracking
mlflow==2.9.2

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Frontend
streamlit==1.28.2

# Visualization
plotly==5.18.0
networkx==3.2.1

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
pymupdf==1.23.26
python-magic-bin==0.4.14  # For Windows
tqdm==4.66.1  # Progress bars